♻️ Refactor: Rename cache service to generalCache
